requirements.txt
## Dependencies
boto3==1.5.15
commis==0.4
confire==0.2.0
Fabric==1.14.0
python-dateutil==2.6.1
python-dotenv==0.7.1
pytz==2017.3
tabulate==0.8.2

## Subdependencies
#asn1crypto==0.24.0
#bcrypt==3.1.4
#botocore==1.8.29
#cffi==1.11.4
#click==6.7
#colorama==0.3.6
#cryptography==2.1.4
#docutils==0.14
#enum34==1.1.6
#futures==3.2.0
#idna==2.6
#ipaddress==1.0.19
#jmespath==0.9.3
#paramiko==2.4.0
#pyasn1==0.4.2
#pycparser==2.18
#PyNaCl==1.2.1
#PyYAML==3.11
#s3transfer==0.1.12

## Python Stuff
#pip==9.0.1
#Python==2.7.13
#setuptools==38.4.0
#six==1.11.0
#wheel==0.30.0
#wsgiref==0.1.2

## Test Dependencies
#apipkg==1.4
#attrs==17.4.0
#coverage==4.4.2
#execnet==1.5.0
#funcsigs==1.0.2
#pluggy==0.6.0
#py==1.5.2
#pyflakes==1.6.0
#pytest-cache==1.0
#pytest-cov==2.5.1
#pytest-flakes==2.0.0
#pytest==3.3.2
